requirements.txt
--find-links=https://girder.github.io/large_image_wheels GDAL
# cartopy
colorcet
folium==0.13.0
geemap[extra]
geopandas
google-cloud-storage
jupyter-server-proxy
keplergl
leafmap
localtileserver
nbserverproxy
os
owslib
palettable
plotly
python-dotenv
requests
streamlit
streamlit-bokeh-events
streamlit-folium
streamlit-keplergl
tropycal
# git+https://github.com/giswqs/leafmap
# git+https://github.com/giswqs/geemap
